[Refactor] BoincHostFinder to be created as Resource [Refactor] BoincManager to be used as resource [Fix] Hosts that are found by the BoincHostFinder are now saved with the correct flag in the BoincManager and the Database [Refactor] Database handling [WPI] Adding/Removing Hosts and projects

+++ jvm/src/main/scala/at/happywetter/boinc/AppConfig.scala
@@ -36,7 +36,7 @@ object AppConfig {
                     secret: String,
                     ssl: SSLConfig)
 
-  case class Host(address: String, port: Short, password: String)
+  case class Host(address: String, port: Int, password: String)
 
   case class Boinc(hosts: Map[String, Host], projects: Projects, connectionPool: Int, encoding: String)
